+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 
 ## Upcoming Release
+* Upgraded Microsoft.Rest.ClientRuntime to 2.3.24
+
+## Version 2.7.5
 * Added `SshCredentialFactory` to support get ssh credential of vm from msal.
 * Fixed the bug of cmdlet fails when -DefaultProfile is set to service principal login context. [#16617]
 * Fixed the issue that authorization does not work in Dogfood environment