+++ src/CognitiveServices/CognitiveServices/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated CognitiveServices PowerShell to use 2022-03-01 version.
+* Added `Get-AzCognitiveServicesAccountModel` cmdlet.
 
 ## Version 1.10.0
 * Updated PowerShell to use 2021-10-01 version.